refactor: breaking up projectdetail into multiple files
